pyproject.toml
[tool.poetry]
name = "sample-api-tests"
version = "0.1.0"
description = ""
authors = ["kpochodyla <kpochodyla@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
pytest = "^7.2.2"
allure-pytest = "^2.13.0"
pydantic = "^1.10.6"
black = "^23.1.0"
isort = "^5.12.0"


[tool.black]
line-length = 88
target-version = ['py310', 'py311']

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
